chore: separate shadow APIs
